move timer into elements directory
