+++ nicegui/elements/timer.js
